+++ lib/recording_studio_oauth.rb
@@ -2,6 +2,8 @@
 
 require "recording_studio_oauth/version"
 require "recording_studio_oauth/configuration"
+require "recording_studio_oauth/protected_resource"
+require "recording_studio_oauth/protected_resource_registry"
 
 module RecordingStudioOauth
   class << self
@@ -12,6 +14,10 @@ module RecordingStudioOauth
     def configure
       yield(configuration) if block_given?
       configuration
+    end
+
+    def protected_resources(api_key: "public")
+      ProtectedResourceRegistry.build(configuration: configuration, api_key: api_key)
     end
   end
 end