+++ test/dummy/config/initializers/recording_studio_oauth.rb
@@ -6,4 +6,7 @@ RecordingStudioOauth.configure do |config|
   config.admin_root_recordable_type_names = ["AdminRoot"]
   config.api_mount_path = "/recording_studio_api"
   config.engine_mount_path = "/recording_studio_oauth"
+  config.mcp_mount_path = "/recording_studio_mcp"
+  config.register_origin_as_protected_resource = false
+  config.extra_protected_resource_paths = []
 end